+++ lib/src/enums/http_exceptions.dart
@@ -1,10 +1,15 @@
 /// The possible exceptions types
 enum HttpExceptions {
+  // Client errors
   ClientError,
+  ExpiredToken,
+  Unauthorized,
+  // Server errors
   ServerError,
+  // Connection errors
   ConnectionError,
   ServerDown,
   ClientOffline,
-  ExpiredToken,
+  // Others
   Other,
 }